+++ maze_input_frame.py
@@ -1,0 +1,95 @@
+from asciimatics.widgets import Frame, Layout, Label, Text, Button
+from asciimatics.exceptions import StopApplication
+import os
+
+
+# Function to get terminal size
+def get_terminal_size():
+    return os.get_terminal_size()
+
+
+def getFixedWidth(width, height):
+    totalSize = width * height
+    return len(str(totalSize))
+
+
+class MazeInputFrame(Frame):
+    def __init__(self, screen, input_data):
+        super(MazeInputFrame, self).__init__(
+            screen,
+            screen.height // 2,
+            screen.width // 2,
+            has_border=True,
+            name="Maze Input Form",
+        )
+
+        self.set_theme("bright")
+
+        # Initialize self.data to store the width and height
+        self.data = {"width": "", "height": ""}
+        self.input_data = (
+            input_data  # Reference to external dictionary to store final input
+        )
+
+        # Layout to organize widgets
+        layout = Layout([1], fill_frame=True)
+        self.add_layout(layout)
+
+        # Add widgets to the layout
+        self.message_label = Label("")
+        layout.add_widget(self.message_label)
+
+        layout.add_widget(Label("Enter maze width (odd number >= 5):"))
+        self.width_input = Text("Width:", "width")
+        layout.add_widget(self.width_input)
+
+        layout.add_widget(Label("Enter maze height (odd number >= 5):"))
+        self.height_input = Text("Height:", "height")
+        layout.add_widget(self.height_input)
+
+        # Button to submit the form
+        layout.add_widget(Button("Submit", self._submit))
+
+        # Finalize layout
+        self.fix()
+
+    def _submit(self):
+        # Save the input data
+        self.save()
+
+        try:
+            # Check if the self.data is None
+            if not self.data:
+                self.message_label.text = (
+                    "Please enter valid integers for width and height!"
+                )
+                self.screen.force_update()
+                return
+
+            # Retrieve width and height and validate them
+            width = int(self.data["width"])
+            height = int(self.data["height"])
+
+            if width < 5:
+                self.message_label.text = "Width must be an odd number >= 5!"
+                self.screen.force_update()
+                return
+
+            if height < 5:
+                self.message_label.text = "Height must be an odd number >= 5!"
+                self.screen.force_update()
+                return
+
+            # If input is valid, store it in the external dictionary
+            self.input_data["width"] = width
+            self.input_data["height"] = height
+            # If input is valid, print the values and exit the form
+            print(f"Maze width: {width}, Maze height: {height}")
+            raise StopApplication("User submitted the form")
+
+        except ValueError:
+            # Handle non-integer input
+            self.message_label.text = (
+                "Please enter valid integers for width and height!"
+            )
+            self.screen.force_update()
